// src/shared/enums/analytics-interval.ts
export enum AnalyticsInterval {
  ANALYTICS_INTERVAL_FIVE_MINUTES = 'five-minutes',
  ANALYTICS_INTERVAL_THIRTY_MINUTES = 'thirty-minutes',
  ANALYTICS_INTERVAL_HOUR = 'hour',
  ANALYTICS_INTERVAL_WEAK = 'weak',
  ANALYTICS_INTERVAL_MONTH = 'month',
  ANALYTICS_INTERVAL_YEAR = 'year',
  ANALYTICS_INTERVAL_GAP = 'gap',
}